// libs/components/list-builder-view-grids/src/lib/modules/grid/types/grid-column-description-model-change.ts
/**
 * @internal
 * @deprecated `SkyGridComponent` and its features are deprecated. We recommend using the data grid instead. For more information, see https://developer.blackbaud.com/skyux/components/data-grid
 */
export interface SkyGridLegacyColumnDescriptionModelChange {
  value?: string;
  id?: string;
  field?: string;
}