// libs/components/list-builder-view-grids/src/lib/modules/grid/types/grid-column-inline-help-popover-model-change.ts
/**
 * @internal
 * @deprecated `SkyGridComponent` and its features are deprecated. We recommend using the data grid instead. For more information, see https://developer.blackbaud.com/skyux/components/data-grid
 */
export interface SkyGridLegacyColumnInlineHelpPopoverModelChange {
  value?: any;
  id?: string;
  field?: string;
}